// src/components/BlogClassification/data.d.ts
export interface articleType {
  /**
   * 类型id
   */
  Id: number;

  /**
   * icon
   */
  icon: string;

  /**
   * 类型名称
   */
  typeName: string;

  /**
   * 类型附属名称
   */
  subTypeName: string;
}

export interface newType {

  /**
   * icon
   */
  icon: string;

  /**
   * 分类名称
   */
  typeName: string;
}